TSI-920 Moved all ECL-related logic to a separate class (EclProcessor) to avoid assembly loading errors if ECL is not installed  (TSI-979)
