+++ source/DD4T.Templates.Base/DD4T.Templates.Base.csproj
@@ -197,7 +197,7 @@
     <Compile Include="Builder\ComponentBuilder.cs" />
     <Compile Include="Builder\ComponentPresentationBuilder.cs" />
     <Compile Include="Builder\ComponentTemplateBuilder.cs" />
-    <Compile Include="Builder\EclFieldsBuilder.cs" />
+    <Compile Include="Utils\EclProcessor.cs" />
     <Compile Include="Builder\FieldBuilder.cs" />
     <Compile Include="Builder\FieldsBuilder.cs" />
     <Compile Include="Builder\KeywordBuilder.cs" />
